+++ src/sections/ui/breadcrumb/Breadcrumb.tsx
@@ -1,6 +1,11 @@
 import { Breadcrumb as BreadcrumbBS } from 'react-bootstrap'
 import { PropsWithChildren } from 'react'
+import { BreadcrumbItem } from './BreadcrumbItem'
 
-export function Breadcrumb({ children }: PropsWithChildren) {
+function Breadcrumb({ children }: PropsWithChildren) {
   return <BreadcrumbBS>{children}</BreadcrumbBS>
 }
+
+Breadcrumb.Item = BreadcrumbItem
+
+export { Breadcrumb }
